backend/requirements.txt
# Django and REST framework
Django==4.2.10
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.1

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Environment variables
python-dotenv==1.0.1

# API documentation
drf-yasg==1.21.7

# File handling
Pillow==10.1.0

# AI integration
openai==1.12.0
anthropic==0.8.1
nltk==3.8.1
scikit-learn==1.3.2

# Task queue
celery==5.3.6
redis==5.0.1

# Testing
pytest==7.4.3
pytest-django==4.7.0

# Production
gunicorn==21.2.0
whitenoise==6.6.0

# Utilities
python-slugify==8.0.1
requests==2.31.0

# Payments
stripe==7.9.0
